requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.0.7
    # via spotipy
requests==2.32.3
    # via spotipy
spotipy==2.24.0
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   requests
    #   spotipy
